requirements.txt
fastapi
httpx
nba_api
uvicorn
pandas
python-multipart
pika
pymongo
python-dotenv